gradle: Replace BundleTaskConvention with BundleTaskExtension
As Gradle are moving to deprecate conventions, we need to react and
replace the use of conventions with extensions. This change moves
the bundle properties into a `bundle` extension.

The existing convention is still in place but delegates to the
extension. In the Gradle 8 timeframe, when Gradle removes convention
support, we will need to remove BundleTaskConvention.
