+++ biz.aQute.bnd.gradle/README.md
@@ -55,10 +55,10 @@ plugins {
 
 This plugin extends the normal [`java` plugin][3] by:
 
-* Extending the `jar` task by adding the `BundleTaskConvention`.
-This adds new properties are to the `jar` task and the task actions are extended to use Bnd to generate a bundle.
+* Extending the `jar` task by adding a `bundle` extension.
+The  `jar` task actions are also extended to use Bnd to generate a bundle.
 The bundle will contain all the content configured into the `jar` task plus whatever additional content is included via the Bnd instructions.
-The `bndfile` property of the task's `BundleTaskConvention` is set to `'bnd.bnd'` if a `'bnd.bnd'` file exists in the project folder.
+The `bndfile` property of the task's `bundle` extension is set to `'bnd.bnd'` if a `'bnd.bnd'` file exists in the project folder.
 * Adds a `baseline` configuration.
 * Adds a `baseline` task of type `Baseline` which baselines the bundle generated by the `jar` task against the prior version of that bundle in the repositories.
 
@@ -72,7 +72,7 @@ plugins {
 
 ## Create a task of the `Bundle` type
 
-The `Bundle` task type is an extension of the [`Jar` task type][18] that adds the `BundleTaskConvention` which adds new properties and uses Bnd to generate a bundle.
+The `Bundle` task type is an extension of the [`Jar` task type][18] that adds a `bundle` extension with properties and uses Bnd to generate a bundle.
 The bundle will contain all the content configured into the task plus whatever additional content is included via the Bnd instructions.
 For example:
 
@@ -107,7 +107,7 @@ You will only need to specify this property if you want to use a different Sourc
 ### classpath
 
 The FileCollection object to use as the classpath for the Bnd builder.
-The default value is _${sourceSet.compileClasspath}_.
+The default value is _${project.sourceSets.main.compileClasspath}_.
 You will only need to specify this property if you want to specify additional classpath elements or replace the classpath.
 
 ### Example
@@ -123,8 +123,11 @@ sourceSets {
 }
 tasks.register('bundle', Bundle) {
   from sourceSets.bundle.output
-  bndfile = 'bundle.bnd'
-  sourceSet = sourceSets.bundle
+  bundle {
+    bndfile = 'bundle.bnd'
+    sourceSet = sourceSets.bundle
+    classpath = sourceSets.bundle.compileClasspath
+  }
 }
 ```
 
@@ -132,17 +135,19 @@ tasks.register('bundle', Bundle) {
 
 The normal way to instruct Bnd on how to build the bundle is to use a bnd file.
 This file will specify the Bnd instructions like `-exportcontents`, etc.
-You can also use the `bnd` property of the `BundleTaskConvention`.
+You can also use the `bnd` property of the `bundle` extension.
 It can be used in several different ways.
 You can configure the `bnd` property with a multiline string:
 
 ```groovy
 jar {
+  bundle {
     bnd '''
 -exportcontents: com.acme.api.*
 -sources: true
 -include: other.bnd
 '''
+  }
 }
 ```
 
@@ -150,9 +155,11 @@ You can also configure the `bnd` property with map notation:
 
 ```groovy
 jar {
-  bnd('-exportcontents': 'com.acme.api.*',
-    '-sources': 'true',
-    '-include': 'other.bnd')
+  bundle {
+    bnd('-exportcontents': 'com.acme.api.*',
+      '-sources': 'true',
+      '-include': 'other.bnd')
+  }
 }
 ```
 
@@ -169,9 +176,9 @@ jar {
 }
 ```
 
-You can even use a combination of the `manifest` property and a `BundleTaskConvention` property.
-But the `BundleTaskConvention` property takes priority over the `manifest` property.
-So if the same instruction is in both places, the one in the `BundleTaskConvention` property will be used and the one in the `manifest` property will be ignored.
+You can even use a combination of the `manifest` property and a `bundle` extension property.
+But the `bundle` extension property takes priority over the `manifest` property.
+So if the same instruction is in both places, the one in the `bundle` extension property will be used and the one in the `manifest` property will be ignored.
 
 ### Replacing use of Gradle's deprecated OSGi Plugin
 
@@ -196,14 +203,15 @@ becomes:
 
 ```groovy
 jar {
-  bnd ('Bundle-Name': 'overwrittenSpecialOsgiName',
-    'Bundle-Vendor': 'MyCompany',
-    'Bundle-Description': 'Platform2: Metrics 2 Measures Framework',
-    'Bundle-DocURL': 'https://www.mycompany.com')
+  bundle {
+    bnd('Bundle-Name': 'overwrittenSpecialOsgiName',
+      'Bundle-Vendor': 'MyCompany',
+      'Bundle-Description': 'Platform2: Metrics 2 Measures Framework',
+      'Bundle-DocURL': 'https://www.mycompany.com')
+  }
 }
 ```
 
-Use the `sourceSet` property to replace use of the `classesDir` property.
 Use of the convenience properties on `OSGiManifest` must be replaced with the specifying the actual OSGi manifest header name in the Bnd instructions.
 For example, `name = 'overwrittenSpecialOsgiName'` is replaced with `bnd 'Bundle-Name: overwrittenSpecialOsgiName'`.
 
@@ -671,7 +679,7 @@ If you do need to write a `build.gradle` file for a Bnd project, there are some 
 * The `bnd.project` property of the project contains the [Project][9] object.
 
 Bnd properties for a project can be accessed in several ways.
-Given the example Bnd property name `foo`, you can use the [bnd extension][16], `bnd.get('foo', 'defaultValue')`, or directly `bnd.foo`.
+Given the example Bnd property name `foo`, you can use the [`bnd` extension][16], `bnd.get('foo', 'defaultValue')`, or directly `bnd.foo`.
 To access Bnd properties without any macro processing you can use the [`unprocessed` function][16], `bnd.unprocessed('foo', 'defaultValue')`.
 
 ## Using other JVM languages with the Bnd Gradle Plugins for Bnd Workspace builds
