+++ biz.aQute.bnd.gradle/src/aQute/bnd/gradle/Baseline.groovy
@@ -242,7 +242,7 @@ public class Baseline extends DefaultTask {
 			if (t instanceof TaskProvider) {
 				t = t.get()
 			}
-			t instanceof Task && t.getConvention().findPlugin(BundleTaskConvention.class) ? t : null
+			t instanceof Task && t.getExtensions().findByName("bundle") ? t : null
 		}
 	}
