+++ biz.aQute.bnd.gradle/src/aQute/bnd/gradle/Bundle.groovy
@@ -10,7 +10,7 @@ import org.gradle.api.tasks.bundling.Jar
  * <p>
  * This task type extends the Jar task type and can be used 
  * for tasks that make bundles. The Bundle task type adds the
- * properties from the BundleTaskConvention.
+ * properties from the BundleTaskExtension.
  *
  * <p>
  * Here is an example of using the Bundle task type:
@@ -20,21 +20,26 @@ import org.gradle.api.tasks.bundling.Jar
  *   description "Build my bundle"
  *   group "build"
  *   from sourceSets.bundle.output
- *   bndfile = project.file("bundle.bnd")
- *   sourceSet = sourceSets.bundle
+ *   bundle {
+ *     bndfile = project.file("bundle.bnd")
+ *     sourceSet = sourceSets.bundle
+ *     classpath = sourceSets.bundle.compileClasspath
+ *   }
  * }
  * </pre>
  */
 public class Bundle extends Jar {
+	private final BundleTaskExtension extension
 	/**
 	 * Create a Bundle task.
 	 *
 	 * <p>
-	 * Also adds the BundleTaskConvention to this task.
+	 * Also adds the BundleTaskExtension to this task.
 	 */
 	public Bundle() {
 		super()
-		getConvention().getPlugins().bundle = new BundleTaskConvention(this)
+		extension = getExtensions().create("bundle", BundleTaskExtension.class, this)
+		getConvention().getPlugins().put("bundle", new BundleTaskConvention(extension, this))
 	}
 
 	/**
@@ -48,7 +53,7 @@ public class Bundle extends Jar {
 	@Override
 	protected void copy() {
 		supercopy()
-		buildBundle()
+		extension.buildBundle()
 	}
 
 	/**
